More efficient API for calling superclass methods
jive.oo.Class supports invoking the overridden implementation of
a method when providing a new implementation in a subclass.  Previously
that was done like this:

    var Subclass = Superclass.extend(function(protect) {
        protect.someMethod = function(a, b) {
            var ret = this._super(a, b);
            return transform(ret);
        };
    });

This patch changes the API to look like this:

    var Subclass = Superclass.extend(function(protect, _super) {
        protect.someMethod = function(a, b) {
            var ret = _super.someMethod.call(this, a, b);
            return transform(ret);
        };
    });

The new API is more verbose; but is a bit more flexible and can
substantially reduce execution time when a class definition is
processed.